requirements.txt
# Agentic AI Career Platform - Python Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
openai>=1.6.0
numpy>=2.0.0
sentence-transformers>=2.2.2
gunicorn==21.2.0
pydantic>=2.5.2
requests>=2.31.0
reportlab>=4.0.0
